requirements.txt
autograd==1.6.2
autograd-gamma==0.5.0
blobfile==2.1.1
certifi==2024.2.2
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
ecos==2.0.13
einops==0.8.0
exceptiongroup==1.2.1
filelock==3.15.4
fonttools==4.53.0
formulaic==1.0.1
fsspec==2024.6.0
ftfy==6.2.0
future==1.0.0
h5py==3.11.0
huggingface-hub==0.23.4
idna==3.7
importlib_resources==6.4.0
iniconfig==2.0.0
interface-meta==1.3.0
joblib==1.4.2
kaplanmeier==0.2.0
kiwisolver==1.4.5
lifelines==0.29.0
lxml==4.9.4
matplotlib==3.9.1
mypy==1.10.0
mypy-extensions==1.0.0
numexpr==2.10.0
numpy==1.26.4
opencv-python==4.10.0.84
openslide-python==1.3.1
osqp==0.6.5
packaging==24.0
pandas==2.2.2
pillow==10.3.0
pip==24.0
pluggy==1.5.0
POT==0.9.3
protobuf==5.26.1
pycryptodomex==3.20.0
pyparsing==3.1.2
pytest==8.2.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
qdldl==0.1.7.post2
regex==2024.5.15
requests==2.32.2
safetensors==0.4.3
scikit-learn==1.3.2
scikit-survival==0.22.2
scipy==1.11.4
seaborn==0.13.2
sentence-transformers==3.0.1
sentencepiece==0.2.0
setuptools==69.5.1
six==1.16.0
tensorboardX==2.6.2.2
threadpoolctl==3.5.0
tokenizers==0.19.1
tomli==2.0.1
torch==1.12.1+cu102
torchvision==0.13.1+cu102
tqdm==4.66.4
transformers==4.42.3
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
wheel==0.43.0
wrapt==1.16.0
yacs==0.1.8
zipp==3.19.2
